requirements.txt
importlib_metadata==8.7.0
jinja2==3.1.6
markdown==3.8
MarkupSafe==3.0.2
piexif==1.1.3
pillow==11.2.1
PyYAML==6.0.2
setuptools==58.0.4
zipp==3.22.0
beautifulsoup4==4.12.2